[FLINK-30464] Exclude metrics during stabilization from scaling decision
So far, the code collects metrics as soon as the job goes into RUNNING mode,
regardless of whether we are still in the stabilization period. This means the
first metric collection window after rescaling will include metrics from the
stabilization period. This is not desirable. The stabilization period is
supposed to allow the job to stabilize. Metrics during stabilization shouldn't
be used to make new scaling decisions.