[FLINK-30575] Ensure zero rates lead to scale down
The logic for computing true processing rates rely on recordsProcessed /
recordProcessingTime style computations which do not really work well when
everything is 0. This leads to no scaling actions when the load suddenly drops
to 0.

The change here takes a pragmatic approach. Instead of special-casing zero
rates throughout the code, we set rates and busyness to near zero values. This
keeps the code functional as-is while ensuring scale down happens.

This concept corresponds to scaling Flink jobs via the parallelism which
naturally cannot be scaled to zero.