[FLINK-31326] Consolidate source scaling logic
Previous changes removed the required for sources to have the `pendingRecords`
metric. This allows to refactor and consolidate the source scaling code. We
remove the option to disable scaling sources because the primary motivation was
to skip scaling legacy sources. The source scaling logic works like this:

* Modern sources (busy time available)

Scale source based on the busy time up. Add extra capacity if the lag
information, provided by the pending records metric, is available. Otherwise
assume 0 lag. Cap by the maximum number of partitions, if available.

* Legacy sources (busy time unavailable)

Leave the source parallelism unchanged, use the current processing rate as the
target rate downstream.