[FLINK-31502] Limit the number of scale operations to reduce cluster churn
Until we move to using the upcoming Rescale API which recycles pods, we need to
be mindful with how many deployments we scale at the same time because each of
them is going to give up all its pods and require the new number of required
pods.

This can cause churn in the cluster and temporary lead to "unallocatable" pods
which triggers the k8s cluster autoscaler to add more cluster nodes. That is
often not desirable because the actual required resources after the scaling have
been settled, are lower.