[FLINK-30575] Limit output ratio in case of near zero values
When the rates drop to zero, we set near zero values to continue to use the same
scaling logic. There is a gap in this logic when only the input rate drops to
zero but the output rate is still measured to be greater than zero. This is
because we do not pass on the output rates directly but the output / input ratio
which becomes very large in this special scenario.

In this change we cap the output rate in case we use near zero input rates. This
has been tested to effectively prevent scale up spikes in pipelines which only
occasionally receive data.