[FLINK-32412] Reduce JobID collision chance
Instead of using Java's hashCode, which is an integer value (32bit), a long
representation of the uid is used. This decreases the chance for an ID
collision drastically:

For a 50% collision chance with random integers, 77000 numbers need to be
generated. For a long value (64 bit) a 50% change of a collision needs 5.1×10^9
random longs. For details look up: "the birthday problem".

A test is added to increase awareness of the problem when changing this part of
the code.