[FLINK-32729] allow starting a flink job in a suspended state.
In our setup we intend migrating flinkDeployments from one cluster to
the other, as part of doing this we want to start the flinkDeploments in
a suspended state with an initialSavePoint configured, and only
configure it to start after we must have stopped flinkDeployment in the
source cluster. This patch was agreed upon in this [slack
thread](https://apache-flink.slack.com/archives/C03GV7L3G2C/p1690825041612599).