[FLINK-32903] Add a load simulation example pipeline for canary testing
The current provided example pipeline generates an infinite amount of load, only
yielding only scaleup decisions. It would be desirable to have a pipeline which
scales up and down in a periodic fashion. Users do not always have enough data
available to do that. Hence, the load needs to be simulated.

The pipeline has been tested and proven to yield very accurate load numbers.