[FLINK-33527] Simplify state store cleanup logic
The state store implementation should take care of optimizing flush
operations. That is already the case for the Kubernetes-based
implementation. The in-memory state store does not require flushing.

The current logic tries to save on API calls but there is no actual savings
because the ConfigMap needs to be retrieved at least once to check whether its
empty. So instead of checking for emptiness, we can simply clear it.

Further, I've removed the optimization to keep some of the data when disabling
the autoscaler. The metric collection contract requires that the entire metric
window contains data. By removing some of the data, once turned on again, we
could add metric gaps to the metric history if there are still metrics left from
the last metric window which overlap with the current. The reason is that the
metric window is counted from the oldest timestamp found in the metric history.