[FLINK-33710] Prevent triggering cluster upgrades for permutations of the same overrides
Previous fix in #720 made the parallelism override string deterministic, but it
will likely result in existing deployments to trigger a one-off unneeded spec
update.

To prevent this and to update only once an actual scaling occurs, we
need to compare the existing overrides with the new ones and check if they are
identical. In this case, we restore the current override string with its own
permutation.