[FLINK-33770] Silence Configuration logging for INFO messages
In #725 we coverted the deprecated keys fallback keys and added test
coverage. This additional change silences logs for fallback keys by setting the
configuration logging to WARN instead of the default INFO.

This avoids spamming the log with unneccessary messages on every reconciliation
loop, while still reporting warnings and errors.