[FLINK-34152] Add an option to scale memory when downscaling (#786)
This adds an option to increase heap memory when removing
TaskManagers. The scaling is applied after adjusting the memory pools (heap,
metaspace, network, managed) and only affects heap memory.

The reason for adding this functionality is that the likelihood of running into
memory constrained scenarios when downscaling is increased after applying memory
tuning. As a precaution, we temporarily increase the memory up to the maximum
allowed TaskManager memory.